fixed problems with webworkers
